Reorganized sequence classes into seq_classes_* structure
New files are "seq_classes_access.cpp", "seq_classes_ref.h", "seq_classes_var.cpp", and "seq_classes_var.h"
