+++ src/mevo_gammas.cpp
@@ -18,7 +18,6 @@
 #include <random>               // gamma_distribution
 
 #include "gemino_types.h"       // integer types
-#include "sequence_classes.h"   // Var* and Ref* classes
 #include "pcg.h"                // pcg seeding
 #include "mevo_gammas.h"        // Gamma* classes
