+++ src/seq_classes_var.cpp
@@ -2,27 +2,261 @@
 /*
  ********************************************************
 
- Methods for adding mutations to sequence classes.
- Note that this file is for just adding the mutations, not for anything related to
- determining where they should go.
+ Basic methods for sequence classes: retrieving info and initializing.
 
  ********************************************************
  */
 
+
 #include <RcppArmadillo.h>
 #include <vector>  // vector class
 #include <string>  // string class
-#include <deque>  // deque class
-#include <algorithm>  // sort
+#include <algorithm>  // lower_bound, sort
+#include <deque>  // deque
 
 
 #include "gemino_types.h"  // integer types
-#include "sequence_classes.h"  // Var* and Ref* classes
-#include "util.h"  // clear_memory
-
+#include "seq_classes_ref.h"  // Ref* classes
+#include "seq_classes_var.h"  // Var* classes
 
 using namespace Rcpp;
 
+
+
+/*
+ ------------------
+ Retrieve a nucleotide (char type) from the variant sequence
+ based on the position in the new, variant sequence
+ ------------------
+ */
+char VarSequence::get_nt(const uint32& new_pos) const {
+    char out;
+    /*
+     Index to the Mutation object nearest to (without being past)
+     an input position (see below for the `get_mut_` fxn):
+     */
+    uint32 mut_i = get_mut_(new_pos);
+    /*
+     If the new_pos is less than the position for the first mutation
+     or if mutations is empty
+     (in which cases `get_mut_` returns `mutations.size()`),
+     we just extract the character from the beginning of the reference string:
+     */
+    if (mut_i == mutations.size()) {
+        out = (*ref_seq)[new_pos];
+        /*
+         If not, then extract the character from the Mutation object that
+         `mut` points to (see below for `get_char_` fxn):
+         */
+    } else {
+        out = get_char_(new_pos, mut_i);
+    }
+
+    return out;
+}
+
+
+
+/*
+ ------------------
+ Retrieve all nucleotides (i.e., the full sequence; std::string type) from
+ the variant sequence
+ ------------------
+ */
+
+std::string VarSequence::get_seq_full() const {
+
+    if (mutations.empty()) return ref_seq->nucleos;
+
+    // Index to the first Mutation object
+    uint32 mut_i = 0;
+
+    std::string out(seq_size, 'x');
+    uint32 pos = 0;
+
+    // Picking up any nucleotides before the first mutation
+    while (pos < mutations[mut_i].new_pos) {
+        out[pos] = (*ref_seq)[pos];
+        ++pos;
+    }
+
+    // Now, for each subsequent mutation except the last, add all nucleotides
+    // at or after its position but before the next one
+    uint32 next_mut_i = mut_i + 1;
+    while (next_mut_i < mutations.size()) {
+        while (pos < mutations[next_mut_i].new_pos) {
+            out[pos] = get_char_(pos, mut_i);
+            ++pos;
+        }
+        ++mut_i;
+        ++next_mut_i;
+    }
+
+    // Now taking care of nucleotides after the last Mutation
+    while (pos < seq_size) {
+        out[pos] = get_char_(pos, mut_i);
+        ++pos;
+    }
+
+    return out;
+}
+
+
+
+
+/*
+ ------------------
+ Retrieve the first part of a sequence from the variant sequence.
+ ------------------
+ */
+std::string VarSequence::get_seq_start(uint32 out_length) const {
+
+    if (out_length > seq_size) out_length = seq_size;
+
+    if (mutations.empty()) return ref_seq->nucleos.substr(0, out_length);
+
+    std::string out(out_length, 'x');
+
+    uint32 mut_i = 0;
+
+    uint32 pos = 0;
+
+    // Picking up any nucleotides before the first mutation
+    while (pos < mutations[mut_i].new_pos) {
+        out[pos] = (*ref_seq)[pos];
+        if (pos == out_length - 1) return out;
+        ++pos;
+    }
+
+    // Now, for each subsequent mutation except the last, add all nucleotides
+    // at or after its position but before the next one
+    uint32 next_mut_i = mut_i + 1;
+    while (next_mut_i < mutations.size()) {
+        while (pos < mutations[next_mut_i].new_pos) {
+            out[pos] = get_char_(pos, mut_i);
+            if (pos == out_length - 1) return out;
+            ++pos;
+        }
+        ++mut_i;
+        ++next_mut_i;
+    }
+
+    // Now taking care of nucleotides after the last Mutation
+    while (pos < seq_size) {
+        out[pos] = get_char_(pos, mut_i);
+        if (pos == out_length - 1) return out;
+        ++pos;
+    }
+
+    return out;
+}
+
+
+/*
+ ------------------
+ Set an input string object to any chunk of a sequence from the variant sequence.
+ Before anything, this function moves `mut` to the location right before this chunk's
+ starting position. I keep this index around so I don't have to iterate through
+ the entire mutation deque multiple times.
+ If end position is beyond the size of the sequence, it changes `chunk_str` to the
+ sequence from the start to the sequence end.
+ If start position is beyond the size of the sequence, it sets `mut` to `mutations.end()`
+ and clears `chunk_str`.
+ ------------------
+ */
+void VarSequence::set_seq_chunk(std::string& chunk_str,
+                                const uint32& start,
+                                const uint32& chunk_size,
+                                uint32& mut_i) const {
+
+    uint32 end = start + chunk_size - 1;
+
+    if (start >= seq_size) {
+        mut_i = mutations.size();
+        chunk_str.clear();
+        return;
+    }
+    // Making sure end doesn't go beyond the sequence bounds
+    if (end >= seq_size) end = seq_size - 1;
+
+    uint32 out_length = end - start + 1;
+
+    // No need to mess around with mutations if there aren't any
+    if (mutations.empty()) {
+        chunk_str = ref_seq->nucleos.substr(start, out_length);
+        return;
+    }
+    // Move mutation to the proper spot
+    while (mut_i < mutations.size()) {
+        if (start < mutations[mut_i].new_pos) break;
+        ++mut_i;
+    }
+    if (mut_i != 0) --mut_i;
+    // Adjust input string size if necessary
+    if (chunk_str.size() != out_length) chunk_str.resize(out_length, 'x');
+
+    uint32 pos = start;
+    uint32 next_mut_i = mut_i + 1;
+
+    /*
+     Picking up any nucleotides before the focal mutation (this should only happen when
+     `mut == mutations.begin()` and `start` is before the first mutation)
+     */
+    while (pos < mutations[mut_i].new_pos && pos <= end) {
+        chunk_str[pos - start] = (*ref_seq)[pos];
+        ++pos;
+    }
+    if (pos > end) return;
+
+    /*
+     Now, for each subsequent mutation except the last, add all nucleotides
+     at or after its position (and `end`) but before the next mutation
+     */
+    while (next_mut_i < mutations.size()) {
+        while (pos < mutations[next_mut_i].new_pos && pos <= end) {
+            chunk_str[pos - start] = get_char_(pos, mut_i);
+            ++pos;
+        }
+        if (pos > end) return;
+        ++mut_i;
+        ++next_mut_i;
+    }
+
+    /*
+     If we reach the last mutation, add nucleotides until `end` (remember that above,
+     I've made sure that `end < seq_size`).
+     */
+    while (pos <= end) {
+        chunk_str[pos - start] = get_char_(pos, mut_i);
+        ++pos;
+    }
+
+    return;
+}
+
+
+/*
+ ------------------
+ Internal function for finding character of either mutation or reference
+ given an index (in the "new", variant sequence) and an index for a
+ single Mutation object.
+ This only works if you've already narrowed it down to the Mutation object
+ that is directly previous to the index position.
+ ------------------
+ */
+char VarSequence::get_char_(const uint32& new_pos,
+                            const uint32& mut_i) const {
+    const Mutation& m(mutations[mut_i]);
+    char out;
+    uint32 ind = new_pos - m.new_pos;
+    if (static_cast<sint32>(ind) > m.size_modifier) {
+        ind += (m.old_pos - m.size_modifier);
+        out = (*ref_seq)[ind];
+    } else {
+        out = m[ind];
+    }
+    return out;
+}
 
 
 
@@ -112,17 +346,17 @@ void VarSequence::add_deletion(const uint32& size_, const uint32& new_pos_) {
 
     /*
      If `mutations` is empty, just add to the beginning and adjust sequence size
-    */
+     */
     if (mutations.empty()) {
         Mutation new_mut(old_pos_, deletion_start, size_mod);
         mutations.push_front(new_mut);
         seq_size += size_mod;
 
-    /*
-     If the mutations deque isn't empty, we may need to edit some of the mutations
-     after this deletion if they're affected by it.
-     See `deletion_blowup_` below for more info.
-     */
+        /*
+         If the mutations deque isn't empty, we may need to edit some of the mutations
+         after this deletion if they're affected by it.
+         See `deletion_blowup_` below for more info.
+         */
     } else {
 
         /*
@@ -156,7 +390,7 @@ void VarSequence::add_deletion(const uint32& size_, const uint32& new_pos_) {
         if (mut_i != 0) {
             --mut_i;
             old_pos_ = deletion_start - mutations[mut_i].new_pos +
-                mutations[mut_i].old_pos - mutations[mut_i].size_modifier;
+            mutations[mut_i].old_pos - mutations[mut_i].size_modifier;
             ++mut_i;
         } else old_pos_ = deletion_start; // (`deletion_start` may have changed)
 
@@ -214,10 +448,10 @@ void VarSequence::add_insertion(const std::string& nucleos_, const uint32& new_p
         mutations[mut_i].size_modifier += size_;
         // Adjust new positions and total sequence size:
         calc_positions(mut_i + 1, size_);
-    /*
-     If `new_pos_` is in the reference sequence following the Mutation, we add
-     a new Mutation object:
-     */
+        /*
+         If `new_pos_` is in the reference sequence following the Mutation, we add
+         a new Mutation object:
+         */
     } else {
         uint32 old_pos_ = ind + (mutations[mut_i].old_pos -
             mutations[mut_i].size_modifier);
@@ -256,7 +490,7 @@ void VarSequence::add_substitution(const char& nucleo, const uint32& new_pos_) {
         // If `new_pos_` is within the mutation sequence:
         if (static_cast<sint32>(ind) <= mutations[mut_i].size_modifier) {
             mutations[mut_i].nucleos[ind] = nucleo;
-        // If `new_pos_` is in the reference sequence following the mutation:
+            // If `new_pos_` is in the reference sequence following the mutation:
         } else {
             uint32 old_pos_ = ind + (mutations[mut_i].old_pos -
                 mutations[mut_i].size_modifier);
@@ -301,13 +535,13 @@ void VarSequence::deletion_blowup_(uint32& mut_i, uint32& deletion_start,
      */
     if (mut_i == mutations.size()) {
         mut_i = 0;
-    /*
-     If it's a substitution and has a position < the deletion starting point,
-     we can simply skip to the next one.
-     If they're equal, we don't iterate.
-     If it's > the deletion starting point, that should never happen if `get_mut_` is
-     working properly, so we return an error.
-     */
+        /*
+         If it's a substitution and has a position < the deletion starting point,
+         we can simply skip to the next one.
+         If they're equal, we don't iterate.
+         If it's > the deletion starting point, that should never happen if `get_mut_` is
+         working properly, so we return an error.
+         */
     } else if (mutations[mut_i].size_modifier == 0) {
         if (mutations[mut_i].new_pos < deletion_start) {
             ++mut_i;
@@ -316,22 +550,22 @@ void VarSequence::deletion_blowup_(uint32& mut_i, uint32& deletion_start,
         } else {
             stop("Index problem in deletion_blowup_");
         }
-    /*
-     If the first Mutation is an insertion, we may have to merge it with
-     this deletion, and this is done with `merge_del_ins_`.
-     This function will iterate to the next Mutation.
-     It also adjusts `size_mod` appropriately.
-     */
+        /*
+         If the first Mutation is an insertion, we may have to merge it with
+         this deletion, and this is done with `merge_del_ins_`.
+         This function will iterate to the next Mutation.
+         It also adjusts `size_mod` appropriately.
+         */
     } else if (mutations[mut_i].size_modifier > 0) {
         merge_del_ins_(mut_i, deletion_start, deletion_end, size_mod);
-    /*
-     If it's a deletion and next to the new deletion, we merge their information
-     before removing the old mutation.
-     (`remove_mutation_` automatically moves the index to the location
-     after the original.)
+        /*
+         If it's a deletion and next to the new deletion, we merge their information
+         before removing the old mutation.
+         (`remove_mutation_` automatically moves the index to the location
+         after the original.)
 
-     If it's not next to the new deletion, we just iterate to the next mutation.
-     */
+         If it's not next to the new deletion, we just iterate to the next mutation.
+         */
     } else {
         if (mutations[mut_i].new_pos == deletion_start) {
             size_mod += mutations[mut_i].size_modifier;
@@ -365,20 +599,20 @@ void VarSequence::deletion_blowup_(uint32& mut_i, uint32& deletion_start,
         // For substitutions, do nothing before iterating
         if (mutations[mut_i].size_modifier == 0) {
             ++mut_i;
-        /*
-         For insertions, run `merge_del_ins_` to make sure that...
+            /*
+             For insertions, run `merge_del_ins_` to make sure that...
              (1) any sequence not overlapping the deletion is kept
              (2) `size_mod` is adjusted properly
              (3) insertions that are entirely overlapped by the deletion are erased
              (4) `mut_i` is moved to the next Mutation
-         */
+             */
         } else if (mutations[mut_i].size_modifier > 0) {
             merge_del_ins_(mut_i, deletion_start, deletion_end, size_mod);
             // as above, stop here if deletion is absorbed
             if (size_mod == 0) return;
-        /*
-         For deletions, merge them with the current one
-         */
+            /*
+             For deletions, merge them with the current one
+             */
         } else {
             size_mod += mutations[mut_i].size_modifier;
             ++mut_i;
@@ -421,10 +655,10 @@ void VarSequence::merge_del_ins_(uint32& insert_i,
      */
     if (deletion_start > insertion_end || deletion_end < insertion_start) {
         ++insert_i;
-    /*
-     Else if the entire insertion is covered by the deletion, adjust size_mod and
-     remove the Mutation object for the insertion:
-     */
+        /*
+         Else if the entire insertion is covered by the deletion, adjust size_mod and
+         remove the Mutation object for the insertion:
+         */
     } else if (deletion_start <= insertion_start && deletion_end >= insertion_end) {
         size_mod += mutations[insert_i].size_modifier; // making it less negative
         /*
@@ -432,10 +666,10 @@ void VarSequence::merge_del_ins_(uint32& insert_i,
          object without us doing anything here.
          */
         remove_mutation_(insert_i);
-    /*
-     Else if there is overlap, adjust the size_mod, remove that part of
-     the inserted sequence, and adjust the insertion's size modifier:
-     */
+        /*
+         Else if there is overlap, adjust the size_mod, remove that part of
+         the inserted sequence, and adjust the insertion's size modifier:
+         */
     } else {
 
         // index for first char to erase from `nucleos`
@@ -545,7 +779,7 @@ uint32 VarSequence::get_mut_(const uint32& new_pos) const {
         stop(
             "new_pos should never be >= the sequence size. "
             "Either re-calculate the sequence size or closely examine new_pos."
-            );
+        );
     }
     /*
      If new_pos is less than the position for the first mutation, we return
@@ -583,3 +817,39 @@ uint32 VarSequence::get_mut_(const uint32& new_pos) const {
 
 
 
+
+
+
+
+
+void VarSet::print() const noexcept {
+
+    uint32 total_muts = 0;
+    for (const VarGenome& vg : variants) {
+        for (const VarSequence& vs : vg.var_genome) {
+            total_muts += vs.mutations.size();
+        }
+    }
+
+    int console_width = get_width();
+
+    int n_spaces = static_cast<int>(
+        std::ceil(static_cast<double>(console_width - 21) / 2)
+    );
+
+    for (int i = 0; i < n_spaces; i++) Rcout << ' ';
+    Rcout << "<< Variants object >>" << std::endl;
+
+    Rcout.imbue(std::locale(""));
+    Rcout << "# Variants: " << size() << std::endl;
+    Rcout << "# Mutations: " << total_muts << std::endl;
+    Rcout << std::endl;
+
+    n_spaces = static_cast<int>(
+        std::ceil(static_cast<double>(console_width - 28) / 2)
+    );
+
+    for (int i = 0; i < n_spaces; i++) Rcout << ' ';
+    Rcout << "<< Reference genome info: >>" << std::endl;
+    reference.print();
+}
